# refactor(tools.component.rollup.config): remove deprecated languages